task: handle arbitrary SCM schemes
Prior to this change we would only parse a task's SCM source if the URL
started with "git://".

In Fedora's Koji they use "git+https://", so this doesn't work.

As long as the urlparse() method can determine any scheme here, we
should treat the value as a valid URL and attempt to parse the .path
attribute.