tests: FakeProxy inherits from Proxy
Now that we're touching proxy.path in _authenticated_path(), in order to
test that, we need our fake proxy's .path to be populated (as bytes, etc)

Instead of implementing all that in the FakeProxy constructor, just
inherit from Twisted's real Proxy class. We're still overriding the
important method with callRemote() so the end result is the same, it's
still a fake, just with more attributes populated so it's closer to the
real thing.