task: support estimating free tasks
We can estimate free tasks by checking if the channel has capacity and
then adding kojid's SLEEPTIME to estimate_duration().

If the channel is at capacity, we'll raise NotImplementedError.
Longer-term, we should evaluate all tasks in the channel and determine
when enough OPEN tasks will complete so we can get to OPEN.