.travis.yml: temporarily install older idna for requests
python-requests pins their idna dependency version, and idna has
released a version newer than the pin declaration. setuptools cannot
figure out what to do, so it bails out.

My workaround is to install an older idna manually. Revert this when the
python-requests authors release a new version that does not pin to
idna<2.8.