estimates: multicall containers in average_build_durations()
With the recent use of multicall in average_build_durations(), we could
no longer handle container packages. The problem is that we have to
estimate those with listBuilds, and technically the listBuilds RPC can
only accept package IDs, not package names:
https://pagure.io/koji/issue/1209

Update average_build_durations() to check input: all the packages must
be containers, or none of them must be containers.

If the packages are not containers, then great, we can do a simple
multicall with getAverageBuildDuration for everything.

If they are containers, then we have to map the packages to package IDs,
and then listBuilds on every package ID, and then average those build
times.