estimates: fix non-container case for average_build_durations()
Fix two bugs when parsing non-container packages in
average_build_durations():

If we have no packages named "*-container", we need to set "containers"
to False so we will take the non-container code path below.

If we're taking the non-container code path, we do not need to call
getPackage, only getAverageBuildDuration. In fact, sending getPackage in
our multicall this screws up the multicall results that we return here.