connection: use trustRoot in _gssapi_login()
Prior to this change, we did not respect the "serverca" profile setting
for Kerberos (GSSAPI) logins. Update the _gssapi_login() method to use
the connection's trustRoot. With this change, treq will validate the
HTTPS server's certficate against the one that we have configured (or
the system bundle if none is configured).