proxy: re-add self.trustRoot assignment
During the cleanup in commit 4cc111c6359977759ed1b1230ef9a6501f5498bf, I
should have simply moved this assignment above the parent constructor
call since it's "not a problem upstream" any more, but instead I
*removed* it altogether. This removal led to an AttributeError, because
one of the main purposes of this class is to use self.trustRoot later on
in callRemote(). Restore this to the proper place here.